chore(server): refactored from restify to express
Osprey will need to serve static content for the ui portion of the app, this is better handled by express than restify.
Future consideration will be to move the UI out of osprey but for now it is much easier and avoids the need to bring
a API gateway into the Kaching system. Will punt on API gateway until absolutely necessary as it adds complexity
that is currently not needed